OnFailureBehavior into compiled_config file
