+++ rust/routee-compass/src/plugin/output/default/eval/mod.rs
@@ -1,12 +1,12 @@
 mod builder;
-mod compiled_expression;
+mod compiled_config;
 mod config;
 mod operation;
 mod plugin;
 
 pub mod ops;
 pub use builder::EvalOutputPluginBuilder;
-pub use compiled_expression::CompiledExpression;
+pub use compiled_config::{CompiledExpression, CompiledOnFailure};
 pub use config::{EvalOutputPluginConfig, ExpressionConfig, NotANumberBehavior, OnFailureBehavior};
 pub use operation::Operation;
 pub use plugin::EvalOutputPlugin;
